Entirely remove rmutil dependency, relocate Vector
