+++ src/execution_plan/execution_plan.c
@@ -10,7 +10,7 @@
 #include "execution_plan.h"
 #include "./ops/ops.h"
 #include "../graph/edge.h"
-#include "../rmutil/vector.h"
+#include "../util/vector.h"
 #include "../query_executor.h"
 #include "../arithmetic/algebraic_expression.h"
 #include "./optimizations/optimizer.h"
