+++ src/execution_plan/ops/op_conditional_traverse.h
@@ -12,7 +12,7 @@
 #include "../../parser/ast.h"
 #include "../../arithmetic/algebraic_expression.h"
 #include "../../GraphBLASExt/tuples_iter.h"
-#include "../../rmutil/vector.h"
+#include "../../util/vector.h"
 
 /* OP Traverse */
 typedef struct {
