+++ src/execution_plan/ops/op_traverse.h
@@ -11,7 +11,7 @@
 #include "op.h"
 #include "../../parser/ast.h"
 #include "../../arithmetic/algebraic_expression.h"
-#include "../../rmutil/vector.h"
+#include "../../util/vector.h"
 #include "../../GraphBLASExt/tuples_iter.h"
 
 /* OP Traverse */
