+++ src/parser/clauses/delete.h
@@ -8,7 +8,7 @@
 #ifndef _CLAUSE_DELETE_H
 #define _CLAUSE_DELETE_H
 
-#include "../../rmutil/vector.h"
+#include "../../util/vector.h"
 #include "../../util/triemap/triemap.h"
 
 typedef struct {
