+++ src/parser/clauses/set.h
@@ -9,7 +9,7 @@
 #define _CLAUSE_SET_H
 
 #include "../ast_common.h"
-#include "../../rmutil/vector.h"
+#include "../../util/vector.h"
 #include "../../util/triemap/triemap.h"
 #include "../ast_arithmetic_expression.h"
