+++ src/query_executor.c
@@ -11,7 +11,7 @@
 #include "graph/node.h"
 #include "graph/graph_type.h"
 #include "stores/store.h"
-#include "rmutil/vector.h"
+#include "util/vector.h"
 #include "parser/grammar.h"
 #include "arithmetic/agg_ctx.h"
 #include "arithmetic/repository.h"
