+++ src/util/vector.c
