+++ src/util/vector.h
